Move Api controller tests to folder #45
Moved Api controller tests to folder to mirror the SUT structure
