+++ AutoRenter.Api.Tests/Controllers/LoginControllerTests.cs
@@ -9,7 +9,7 @@ using Microsoft.AspNetCore.Mvc;
 using Moq;
 using Xunit;
 
-namespace AutoRenter.Api.Tests
+namespace AutoRenter.Api.Tests.Controllers
 {
     public class LoginControllerTests
     {
