+++ AutoRenter.Api.Tests/Controllers/SkusControllerTests.cs
@@ -9,7 +9,7 @@ using AutoRenter.Api.Tests.Helpers;
 using AutoRenter.Domain.Interfaces;
 using AutoRenter.Domain.Models;
 
-namespace AutoRenter.Api.Tests
+namespace AutoRenter.Api.Tests.Controllers
 {
     public class SkusControllerTests
     {
